requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
aerich
aiomysql
python-dotenv
pydantic[email]
